% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corHist.R
\name{corHist}
\alias{corHist}
\title{Histogramm of correlations}
\usage{
corHist(data, var1, var2, cluster)
}
\arguments{
\item{data}{The data set to be used. Should consist of at least two numeric variables and one cluster variable.}

\item{var1}{The first varaible to correlate.}

\item{var2}{The second variable to correlate.}

\item{cluster}{The level 2 cluster variable, in which both the \code{var1} and \code{var2} are nested.}
}
\description{
Plots a histogram of correlations to make Simpson's paradox visible.
}
\details{
A histogram of correlations, [[1]] group means for the two level-1
variables and [[2]] their correlation on level-2. Currently,
the pearson correlation is used for computing the correlation.
}
\seealso{
\code{\link[stats]{cor}}
}
