% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMiss.R
\name{findMissing}
\alias{findMissing}
\title{Find missing or invalid data}
\usage{
findMissing(data, values = c(NA), invalids = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{values}{A vector of values to look or not to look for (see
\code{invalids}). Defaults to \code{NA}.}

\item{invalids}{logical. Are the \code{values} entered valid or invalid
values? See Details.}
}
\value{
Returns a list with an entry for each variable that has missing
 values (or invalid values), the position of the missing or invalid
 values as well as the percentage of missing or invalid values per
 variable. Add some values here.
}
\description{
A function to find missings values or values that should not be defined
in a data frame.
}
\details{
If \code{invalids} is set to \code{FALSE}, the values entered in
\code{values} are the valid ones, all other values are reported. If set
to \code{TRUE}, values specified in \code{values} are invalid and
reported.
}
