% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{Get correlation tables for data frames.}
\usage{
getTable(data, apa = F, xlsx = F, xlsxName = "Result_getTable",
  diag = "", diag2 = "", round = 3, detachPackages = FALSE, ...)
}
\arguments{
\item{data}{The data frame for which the correlations should be
computed. Non-numerics are excluded from further processing}

\item{apa}{logical, apa Should APA style be approximated, leading to a
table with an empty upper half. Default is to \code{FALSE}.}

\item{xlsx}{logical, shoudl an xlsx-file be produced? Default is to
\code{FALSE}.}

\item{xlsxName}{The name of the resulting xlsx-file, if \code{xlsx =
TRUE}. Default is to \code{Result getTable}.}

\item{diag}{The value that should be displayed in the first diagonal of
the resulting table, e. g. \code{mean}. Default is empty.}

\item{diag2}{The value that should be displayed in the second diagonal
of the resulting table, e. g. \code{var}. Default is empty.}

\item{round}{The number of digits that should be displayed. Is passed on
to \code{round}. Default is \code{3}.}

\item{...}{Other parameters passed on to \code{pearson.cor}.}
}
\value{
Returns a data frame with correlations. See Details.
}
\description{
A function to generate correlation tables that can be easily transformed
to APA tables.
}
\details{
This is a function that returns a table of correlations and, depending
on \code{apa}, a table that is losely formatted as an apa table. Pearson
correlations (using \code{cor} of all variables are computed. Their
signifcance levels are also reported, based on the function
\code{cor.test} from the  \code{Hmisc}-package.
}
\seealso{
\code{\link{cor}} for the functon that computes the
  correlation, \code{\link[Hmisc]{cor.test}} for the function that
  computes the \emph{p}-values, and \code{\link[xlsx]{write.xlsx}} for
  the function that writes the Exceltable.
}

