% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimme_list.R
\name{gimme_list}
\alias{gimme_list}
\title{Turns Long Data Into a List of Data Frames}
\usage{
gimme_list(data, state_vars, id_var, min_Ti)
}
\arguments{
\item{data}{The data in long format}

\item{state_vars}{The state_variables (level 1) that should be extracted from the data set and kept for the list. Must not contain the ID-variable}

\item{id_var}{The ID/group variable that identifies participants on level-2}

\item{min_Ti}{The minimum number of measurement occasions required for each participant Participants with less measurement occasions will be dropped}
}
\value{
Returns a list of data frames as required by \code{\link[gimme]{gimmeSEM}}. Each entry in the list is a data frame for one participant/group.
}
\description{
The function turns wide data into a list of data frames that can be used as input for the command \code{\link[gimme]{gimmeSEM}}.
}
\examples{
## simulate data
data_long <- data.frame(ID = sort(rep(letters[1:10], 3)),
                       var1 = sort(rep(1:10, 3)),
                       var2 = sort(rep(11:20, 3)))
gimme_list(data_long, state_vars = c("var1", "var2"), id_var = "ID", min_Ti = 2)
}

