% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_mod.R
\name{lav_fit}
\alias{lav_fit}
\title{Get fitmeasures from lavaan objects as text.}
\usage{
lav_fit(fit, round = 2, extract.in = "word")
}
\arguments{
\item{fit}{A fitted lavaan object.}

\item{round}{The number of digits to round to.}

\item{extract.in}{The format to extract in, only form is currently word.}
}
\value{
The function returns a text with the Chisq-value, the corresponding \emph{df} and \emph{p}-value, the CFI, RMSEA and SRMR. Best used for objects fitted with \code{estimator = "MLR"}.
}
\description{
This function extracts the most common used fit measures in a text format, such that they are easy to extract and copy/edit in a word document. \strong{The default extracts are the robust fitmeasures}.
}
\details{
Further implementations:

- specify fit.measures to extract
 - extract in order to plot under a plot
 - extract as latex
 - extract as markdown
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '
fit <- lavaan::cfa(HS.model, data=lavaan::HolzingerSwineford1939, estimator = "MLR")
lav_fit(fit, 2)
}
