% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSimple.R
\name{lmerSimple}
\alias{lmerSimple}
\title{lmerSimple}
\usage{
lmerSimple(criterion, fixed, random, interaction, cluster, data)
}
\arguments{
\item{criterion}{The value that is predicted.}

\item{fixed}{The fixed part of the multilevel-model. Can be variables on any level.}

\item{random}{The random part of the multilevel-model. Can be variables on any level.}

\item{interaction}{The interaction variables. Please enter as interaction term in \code{lmer} format.}

\item{cluster}{The cluster variable in the data frame.}

\item{data}{The data set.}
}
\value{
Returns an \code{lmer} model.
}
\description{
A wrapper function for \code{lmer} models. The input are only character vectors. the function can therefore be run over lists of character vectors.
}
\details{
The model assumes always random intercepts.
}
\examples{
library(lme4)
fit <- lmerSimple("Reaction", "Days", "Days", "", "Subject", sleepstudy)
fit
}

