% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSimple.R
\name{lmertvalue}
\alias{lmertvalue}
\title{Get t-values of multilevel-models}
\usage{
lmertvalue(fit)
}
\arguments{
\item{fit}{A fitted \code{lmer} model.}
}
\value{
Returns a data frame with t-values of the fixed effects.
}
\description{
Get t-values of multilevel-models
}
\examples{
fit <- lmerSimple("Reaction", "Days", "Days", "", "Subject", sleepstudy)
fit
}

