% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bold.R
\name{make_bold}
\alias{make_bold}
\title{Make Entries in Data Frames Bold for Markdown Table}
\usage{
make_bold(data.frame, cut = 0.3)
}
\arguments{
\item{data.frame}{A data frame. Numeric entries will be made bold.}

\item{cut}{The cutoff, numeric. (Absolute) values that are larger than cur will be transformed.}
}
\value{
The same data frame, but with entries formatted to be bold
}
\description{
Make Entries in Data Frames Bold for Markdown Table
}
\examples{
library(psych)
# from psych::fa example:
Thurstone.33 <- as.matrix(Thurstone.33)
mle3 <- fa(Thurstone.33,3 ,rotate="none",fm="mle")
fs <- psych::fa.sort(mle3)
loadings <- round(as.data.frame(fs$loadings[1:nrow(fs$loadings), 1:ncol(fs$loadings)]), 2)
make_bold(loadings)
}

