% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machtItems.R
\name{matchItems}
\alias{matchItems}
\title{Match short and long item names in EFA}
\usage{
matchItems(efa.result, codebook, shortitem = "Namen", longitem = "Items",
  digits = 3, cut = NULL)
}
\arguments{
\item{efa.result}{A fitted EFA from \code{\link[psych]{efa}}.}

\item{codebook}{The codebook, in which both the \code{shortitem} and \code{longitem} are to be found.}

\item{shortitem}{The variable in the codebook containing the short names of the item, which were also used in the data frame for the \code{\link[psych]{efa}}.}

\item{longitem}{The variable in the codebook containing the full item (which was for example given to the participant in the study).}

\item{digits}{The number of digits that should be displayed in the resulting data frame.}

\item{cut}{The smalles loadings that should be displayed. If \code{NULL}, all loadings are returned.}
}
\value{
A data frame with the \code{shortitem} as rownames, in the first column the long items (from \code{longitem}), and in the next columns the factors/components that were extracted during the EFA, witht heir factor loadings.
}
\description{
Get long items from an item-codebook and the results of an exploratory factor analysis (EFA) combined in one data frame.
}
\details{
A common problem is the match of short item names, which are used in a data frame, to their rather long counterparts, i. e. the full item. This function allows matching full items (such as "I like to run an efa") with their short counterpart which was used in the data frame (such as "efa_like"). The matching is based on an item codebook.
}
\seealso{
\code{\link[psych]{efa}} to run an EFA, \code{\link[openxlsx]{write.xlsx}} to write the resulting data frame into a user friendly .xlsx format.
}

