% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSimple.R
\name{merge.list.t}
\alias{merge.list.t}
\title{Make a data frame of t-values.}
\usage{
\method{merge}{list.t}(list)
}
\arguments{
\item{A}{list of t-values.}
}
\value{
A data frame with t-values.
}
\description{
Make a data frame of t-values.
}

