% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Modus of a variable}
\usage{
modus(variable, useNA = "no", equal.dist = 99)
}
\arguments{
\item{variable}{The variable containing}

\item{useNA}{Should NAs be used? Will be passed on to \code{\link{table}}. Can be either \code{"no"}, \code{"ifany"}, or \code{"always"}.}

\item{equal.dist}{If all entries are equally often, this value will be returned as output. Default is to 99.}
}
\value{
The modus will be returned.
}
\description{
Get the modus of a variable, i. e. the value that is most frequent in an variable (can be numeric and character)
}
\details{
If two (not all) values are equally often in the variable, the first value of these will be returned. A message is issued.
}
\examples{
letter_samp <- sample(letters, 50, TRUE)
modus(letter_samp)
##
set.seed(9)
m <- matrix(sample(1:5, 30, replace=TRUE), 10, 3, byrow=TRUE)
m[c(23, 12)] <- NA
m[c(4)] <- 1 # no variance
data <- as.data.frame(m)
apply(data, 1, modus)
##
# message, since 2 and 3 are equally often.
modus(c(2, 2, 2, 3, 3, 3, 4, 4))
}

