% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedAlpha.R
\name{nestedAlpha}
\alias{nestedAlpha}
\title{Compute Alpha for Nested Data Structures}
\usage{
nestedAlpha(item.level.1, level.2, level.3, data, ...)
}
\arguments{
\item{item.level.1}{The variable that includes the items on level 1.}

\item{level.2}{The variable that includes the nests on level 2, for example days or measurement occasions in which the items are nested.}

\item{level.3}{The cluster in which the days are clustered, usually persons, groups, classes.}

\item{data}{The data frame that contains all of the above variables.}

\item{...}{Other parameters that can be passed on to \code{\link[lme4]{lmer}}.}
}
\value{
An alpha for nested data.
}
\description{
This function returns a nested alpha for multilevel data as proposed by Nezlek (2016).
}
\details{
Nezlek, 2016, describes one way in that the internal consistency as a measure for reliabilty can be obtained for longitudinal data. The data structure for these types of examinations usually is items (for example all items of one scale) nested in days (for example 10 days of a diary study), which are again nested in persons (all subjects).

To estimate the reliability, one needs to consider the nested structure.



Please consider the following points before running the analysis:
\itemize{
\item {Coding of items} As is also the case with Alpha, all items must be coded in the direction of the construct.
\item {All items are weighted equally}
\item {All items are assumed to be parallel (i. e. equal error-variances, equal loadings, and unidimensional)}
\item {The variable on level-1 should not be standardized or transformed, but the raw scores}
}

The function relies on the \code{\link[lme4]{lmer}} function in order to run the multilevel model. The multilevel model is run with standard settings, however, these can be changed using the \code{...}-argument.
}
\examples{
sample data frame, 10 persons, 10 days with 4 items per day
person <- sort(rep(letters[1:10], 40))
day <- rep(sort(rep(seq(1:10), 4)), 10)
variable <- rnorm(400, 50, 10)
data <- data.frame(person, day, variable, stringsAsFactors = FALSE)

nestedAlpha(item.level.1 = "variable",
            level.2      = "day",
            level.3      = "person",
            data         = data)
}
\author{
K. T. Horstmann \email{horstmak@hu-berlin.de}
}
\references{
Nezlek, J. B. (in press). A practical guide to understanding reliability in studies of within-person variability. \emph{Journal of Research in Personality}. \url{http://dx.doi.org/10.1016/j.jrp.2016.06.020}.
}

