% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omegaW.R
\name{omegaW}
\alias{omegaW}
\title{OmegaW for fitted lavaan objects}
\usage{
omegaW(fit)
}
\arguments{
\item{fit}{A fitted lavaan object.}
}
\description{
Compute factor reliability Omega_W for latent factors
}
\details{
The function computes omega_w, based on standardized factor loadings.
}
\seealso{
\code{\link[lavaan]{cfa}}
}

