% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameCodebook.R
\name{renameCodebook}
\alias{renameCodebook}
\title{A command to match variable names in a given data set with variable names from a codebook.}
\usage{
renameCodebook(data, codebook, code.old.vars = "varname",
  code.new.vars = "new.varname")
}
\arguments{
\item{data}{The data frame.}

\item{codebook}{The codebook which must be in the current working directory as a data frame}

\item{code.old.vars}{The column in the codebook, in which the original (short) variable names are stored (default is to \code{varname}, as in Unipark)}

\item{code.new.vars}{The column in the codebook, in which the new (usually longer, meaningful) variable names are stored (default is to \code{new.varname}, as in Unipark)}
}
\value{
Returns the original data set with names replaced from the Codebook. All names in the data set that do not appear in the Codebook remain unchanged.
}
\description{
A data set contains variable names, which are usually short. The more meaningful variable names are stored in a codebook, e. g. in Excel. This function allows matching and replacing/renaming the variable names in the data set with the variable names from the codebook.
}
\details{
Details to be shown
}
\seealso{
Not much
}

