% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsonCor.R
\name{simpsonCor}
\alias{simpsonCor}
\title{Detect simpsons paradox in multilevel data}
\usage{
simpsonCor(data, outcome, scales, cluster, ...)
}
\arguments{
\item{data}{The data frame to be used.}

\item{outcome}{The dependend variable, i. e. a variable to be used as a criterion in multilevel analysis.}

\item{scales}{The set of variables which are all correlated with \code{outcome} (and not amongst each other).}

\item{cluster}{The level 2 cluster variable, in which both the \code{var1} and \code{var2} are nested.}

\item{...}{Other arguments passed on to \code{\link{cor}}.}
}
\value{
A plot showing level 1 and level 2 correlations plotted against each other.
}
\description{
Allows visual inspection of multilevel data and the relations between a set of variables and one outcome variable, both on level 1 and on level 2.
}
\details{
The function computes all correlations between a set of variables and one outcome variable on level-1 (within cluster) and level-2 (between cluster, i. e. between the cluster means). The correlations on different levels are then plotted against each other. The more a point is off the diagonal, the more the correlations differ on level 1 and level 2.
}
\seealso{
\code{\link{cor}} for the function that computes the correlation.
}

