% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_mod.R
\name{sort_mod}
\alias{sort_mod}
\title{Sorted Modification Indices from lavaan objects}
\usage{
sort_mod(fit, cut = 10)
}
\arguments{
\item{fit}{A fitted lavaan object.}

\item{cut}{A cut.off for modification indices to display. The default value is 10.}
}
\value{
A data frame with sorted modification indices for any fitted lavaan object
}
\description{
Sorted Modification Indices from lavaan objects
}
\examples{
HS.model <- ' visual  =~ x1 + x2 + x3
textual =~ x4 + x5 + x6
speed   =~ x7 + x8 + x9 '
fit <- lavaan::cfa(HS.model, data=lavaan::HolzingerSwineford1939, estimator = "MLR")
sort_mod(fit, cut = 10)
}

