% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tStats.R
\name{tStats}
\alias{tStats}
\title{tStats as text}
\usage{
tStats(t.fit, round = 2, out.text = "plain")
}
\arguments{
\item{t.fit}{The object containing the result of \code{t.test}}

\item{round}{Number of decimals to which the test-statistic is rounded.}

\item{out.text}{The format of the output text. Currently available are
\itemize{
\item {\code{"plain"}} for plain text output (needs only manual editing for word, for example)
\item {\code{"plot"}} for an output that can be used as input for example in \code{text}.
\item {\code{"latex"}} will follow for easy use in markdown.
}}
}
\value{
Returns a character string with the \emph{t}-statistics of a t-test.
}
\description{
Get the \emph{t}-statistics from the object of \code{t.test}.
}
\examples{
fit <- t.test(1:10, y = c(7:20))
boxplot(1:10, c(7:20))
mtext(side = 1, line = 2, tStats(fit, out = "plain"))



}
\author{
Kai T. Horstmann
}

