% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_recode.R
\name{valid_recode}
\alias{valid_recode}
\title{Recode values to 0 and 1 according to predefined vector of valid elements}
\usage{
valid_recode(variable, valid_values, keep.na = FALSE)
}
\arguments{
\item{variable}{The variable that should be recoded to 0 and 1, must be a vector.}

\item{valid_values}{The vector of values that are valid, i. e. will be returned as 1.}

\item{keep.na}{Logical, should missings be kept as \code{NA}? Default is to \code{FALSE}, returning 0 for a missing value.}
}
\value{
A vector containing 0 and 1 (and \code{NA} if specified) of type \code{integer}.
}
\description{
The function recodes a vector of elelemnts into 0 and 1, based on a vector of valid values. It assumes that any correct answer in a questionnaire will be coded as 1 and all other answers (per default, even \code{NA}) will be returned as 0.
}
\examples{

data <- data.frame(var1 = c("7", NA, "five", "nine", "eight", "six",
                   "x", "X", "100", "50", " five", "fife",
                   "5", " Five"),
                   var2 = c(1, 2, 3, 4, 1, NA, 9, 1, 14, 15, 0, 1, 3, 6),
                   id = c(1:14),
                   stringsAsFactors = FALSE)
valid_recode(variable = data$var1,
                           valid_values = c("five", "Five", "5"))
valid_recode(variable = data$var2, valid_values =  c(1, 2), keep.na = TRUE)
}

