% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{vlookup}
\usage{
vlookup(data, look.for, look.in, return.from, return.multiple = FALSE,
  return.na = TRUE)
}
\arguments{
\item{data}{A data frame in which the values should be looked up}

\item{look.for}{The values to look for in the column \code{look.in} in the data frame.}

\item{look.in}{The column to which the values from \code{look.for} are matched.}

\item{return.from}{The column from which the corresponding values are returned.}

\item{return.multiple}{If one value from \code{look.for} would return multiple values from \code{return.from}, should these multiple values be returned? Default is to FALSE.If set to FALSE and multiple matches occur, an error is returned.}

\item{return.na}{Should missing values be returned? Defaul is to TRUE.}
}
\value{
A vector with the corresponding values from the column \code{return.from}. Length must not be the same as \code{look.for}, but usually is if \code{return.multiple = FALSE} and \code{return.na = TRUE} and all values from \code{look.for} are in \code{look.in}.
}
\description{
This function is an implementation of the vlookp function from Excel. It is used to return values from a column in a data frame, which correspond to certain values in another column in the same data frame. These other values are taken as input for the function.
}
\examples{
data <- data.frame(a = c(1, 2, 3, 5, 5, 3, NA, 4, 4),
b = c("A", "B", "C", "D", "D", "E", "F", "D", NA),
stringsAsFactors = TRUE)
vlookup(data, look.for = c(1, 5), look.in = "a", return.from = "b", return.multiple=TRUE, return.na = TRUE)
}

