% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerSimple.R
\name{plot.tvalues}
\alias{plot.tvalues}
\title{Plot t-values from \code{lmer}-models}
\usage{
\method{plot}{tvalues}(t.value.frame)
}
\arguments{
\item{t.value.frame}{A data frame that contains t-values of the fixed effects. Usually an object returned from \code{merge.list.t}}
}
\value{
Plots t-values and fixed effects.
}
\description{
Plot t-values from \code{lmer}-models
}

